Fixed warning in sample.

+++ linux-sample/example.c
@@ -114,7 +114,7 @@ static kii_bool_t state_handler(
         if ((*writer)(kii, "{\"power\":") == KII_FALSE) {
             return KII_FALSE;
         }
-        if ((*writer)(kii, air_conditioner.power == KII_JSON_TRUE
+        if ((*writer)(kii, (int)air_conditioner.power == (int)KII_JSON_TRUE
                         ? "true," : "false,") == KII_FALSE) {
             return KII_FALSE;
         }
